Move product pages to use dynamic routes and BC SF API
